Add files via upload
If you choose to compile from source using Xcode be sure to unzip "orbital.qtz.zip."  This was only compressed to fall under the 25MB upload limit.  There's a high chance that your build will fail but go for it.  This code is from 2010 and although it compiled without error then (hence the precompiled application) It has not be maintained since then.